split the tls mod
